MA-6829 update bootloader when do OTA
Add two function to update bootloader PackageExtractBootloaderFn WriteRawBootloaderFn.
PackageExtractBootloaderFn is used to update ext4.
	SD: /dev/block/mmcblkx  offset:1024
	emmc: /dev/block/mmcblkxboot0 offset:1024.
	      emmc bootloader partition is read only default.
WriteRawBootloaderFn. is used to update ubifs.
	nand: /dev/mtd/mtd1 the offset is based on nand chip
	firmware2 = firmware1 + (P-8*B)/2 (P:partition size B:Block size)
	nand chip MT29F64G08AFAAA's offset is 36M.

[Android_MX6QP_ARD] NAND OTA over ADB: Error alerts that 'E: Cannot load volume /misc' before and after updating OTA package.
There is not misc partition when nand boot.
This log is showed on screen in recovery mode, So need to delete it.

[Android_MX7D/6SL_SD/EVK] OTA over ADB: Returned result was [t] after updating package, better to be replaced with more recognizable string.
The value of result is according the last one command of ota script.
If the last command is unmount("/system"), The resule will be [/system]
If the last command is package_extract_bootloader\package_extract_file, The resule will be [t].
The value of resule do not make sense.

[Android_6QP/SX_ARD]OTA: U-boot time can't update after finish NAND OTA.
The system boot from firmware2 default on sabreauto nand in past time.But it boot from firmware1 now.
The rule is firmware1 is the default boot, if firmware1 failed, it will boot from firmware2.
For OTA upgrading,we need update firmware1 and firmware2.

Coverity 29225: CHECKED_RETURN in Component recovery
Fix the CHECKED_RETURN for lseek.

This commit merge with below commits:
1e45265 MA-6818 [Android_MX6QP_ARD] NAND OTA over ADB:
2008d03 MA-6866 [Android_MX7D/6SL_SD/EVK] OTA over ADB:
0121839 MA-7301 [Android_6QP/SX_ARD]OTA: U-boot time can't updat
a0ac701 MA-7960 Coverity 29225: CHECKED_RETURN in Component recovery

Change-Id: Iadea79ae74681cc3fc80c1b3ea754e13a8cbcbdb
Signed-off-by: zhang sanshan <b51434@freescale.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/487
Reviewed-by: zhang sanshan <sanshan.zhang@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/831
Reviewed-by: Xiaowen Liu <xiaowen.liu@nxp.com>